Remove explicit dependency on oidc from tests
IdentityServer depends on version 8.0.3 of the oidc auth handler, while
we only use 8.0.1. This is normally fine, but if we explicitly take a
dependency on both the handler at version 8.0.1 and identity server,
then our explicit dependency is a downgrade of what identity server
wants, producing a warning. We don't actually need the explicit
dependency, and removing it fixes the build.